Test: Skip Kube-dns if the Kubernetes version is 1.11
[ upstream commit 8c8f7ae7df1e505883ceb7fd1bc1ec6fa63e264c ]

On Kubernetes 1.11 the dns engine is Coredns instead of Kube-dns. With
this change we avoid to have two DNS services installed when the
Coredns is in place.

+++ test/provision/k8s_install.sh
@@ -141,9 +141,10 @@ if [[ "${HOST}" == "k8s1" ]]; then
     kubectl taint nodes --all node-role.kubernetes.io/master-
 
     sudo systemctl start etcd
-
-    kubectl -n kube-system delete svc,deployment,sa,cm kube-dns || true
-    kubectl -n kube-system apply -f ${PROVISIONSRC}/manifest/dns_deployment.yaml
+    if [[ $INSTALL_KUBEDNS -eq 1 ]]; then
+        kubectl -n kube-system delete svc,deployment,sa,cm kube-dns || true
+        kubectl -n kube-system apply -f ${PROVISIONSRC}/manifest/dns_deployment.yaml
+    fi
 
     $PROVISIONSRC/compile.sh
 else
